MR-2669 - added mixins
